CSS transitions on the left section
Added JS with comments so that the "card-flip" is automatically initiated after a second.

Added CSS with comments so that the animation/keyframe the svg left section moves ups, down, and then back to the middle.

Removed the testing text on the left section on both divs.

Removed the duck.svg.

Changed fill color on mandala.svg to a dark blue